refactor: extract ComponentDetailRow and fix multi-component double-counting
Address PR review feedback:
- Extract ComponentDetailRow.vue sub-component to eliminate ~150 lines of
  duplicated template between At Risk and On Track blocks
- Fix release-level forecast double-counting: issues belonging to multiple
  components were counted once per component when summing remaining work;
  now deduplicates by issue key before computing totalRemaining
- Remove unused styling helpers from ProductReleaseCard (moved to sub-component)

Made-with: Cursor
